% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/ensembl.R
\name{EnsemblGenomes}
\alias{EnsemblGenomes}
\alias{list_ensembl_taxa}
\alias{list_ensembl_releases}
\alias{list_ensembl_species}
\alias{list_ensembl_files}
\title{EnsemblGenomes}
\usage{
list_ensembl_taxa()

list_ensembl_releases(taxon)

list_ensembl_species(
  taxon,
  release = utils::head(list_ensembl_releases(taxon)$release, 1)
)

list_ensembl_files(species, taxon = NULL, release = NULL)
}
\arguments{
\item{taxon}{any of \code{list_ensembl_taxa()$taxon}}

\item{release}{any of \code{list_ensembl_releases(taxon)$release}}

\item{species}{any of \code{list_ensembl_species(taxon, release)}}
}
\value{
A tibble listing available taxa, Ensembl releases and supported
species.
}
\description{
EnsemblGenomes provides access to genome references and
annotations from Ensembl
}
\details{
\subsection{List of Ensembl species}{

As of Jul. 2025, there are around:
\itemize{
\item vertebrate: ~ 325 species [release-112]
\item metazoa: ~ 335 species [release-59]
\item protists: ~ 237 species [release-59]
\item fungi: ~ 1502 species [release-59]
\item plants: ~ 153 species [release-59]
\item bacteria: ~ 31332 species [release-59]
}
}
}
\examples{
list_ensembl_taxa()
list_ensembl_releases('vertebrate')
list_ensembl_releases('fungi')
list_ensembl_species('vertebrate', release = 'release-110')
list_ensembl_files("amphiprion_percula")
list_ensembl_files("amphiprion_percula", release = 'release-99')
}
